+++ Source/TDMultipartDownloader.m
@@ -91,7 +91,8 @@
 
 
 - (void)connectionDidFinishLoading:(NSURLConnection *)connection {
-    LogTo(SyncVerbose, @"%@: Finished loading (%u attachments)", self, _reader.attachmentCount);
+    LogTo(SyncVerbose, @"%@: Finished loading (%u attachments)",
+          self, (unsigned)_reader.attachmentCount);
     if (![_reader finish]) {
         [self cancelWithStatus: _reader.status];
         return;
@@ -139,7 +140,7 @@ TestCase(TDMultipartDownloader) {
              CAssert(writer);
              CAssert([writer install]);
              NSData* blob = [db.attachmentStore blobForKey: writer.blobKey];
-             Log(@"Found %u bytes of data for attachment %@", blob.length, attachment);
+             Log(@"Found %u bytes of data for attachment %@", (unsigned)blob.length, attachment);
              NSNumber* lengthObj = [attachment objectForKey: @"encoded_length"] ?: [attachment objectForKey: @"length"];
              CAssertEq(blob.length, [lengthObj unsignedLongLongValue]);
              CAssertEq(writer.length, blob.length);
